Moved action menu back into Monitor page as separate component, refactored modal
